refactor(leaderboard pages): move files and standardise methods
