+++ src/components/leaderboard/LeaderboardChild.tsx
@@ -1,37 +1,29 @@
 "use client";
-import { LeaderboardRankingsInterface } from "@/app/leaderboard/page";
 import Image from "next/image";
 import { useState } from "react";
+import { LeaderboardRankingsInterface } from "./LeaderboardAsync";
 
-export function LeaderboardSkeleton() {
-	return (
-		<li className="animate-pulse transition duration-300 ease-in-out dark:bg-zinc-900 bg-zinc-300 flex flex-wrap items-center w-full my-4 p-2 px-4 rounded-lg group hover:bg-zinc-700 dark:hover:bg-zinc-800 shadow overflow-hidden">
-			<div className="flex-shrink-0 w-72 truncate dark:text-zinc-700 text-zinc-100 text-left">
-				<div style={{ textIndent: "100%" }} className="w-16 h-16 rounded-full inline-block mx-4 bg-gray-400"></div>
-			</div>
-
-			<div className="flex-grow p-4 w-full md:w-auto overflow-hidden">
-				<div className="transition duration-1000 ease-in-out dark:text-white text-black text-left opacity-0 group-hover:opacity-100 overflow-hidden"></div>
-				<div className="transition duration-300 ease-in-out mt-1 mb-1 w-full h-1 bg-zinc-700 group-hover:bg-zinc-500 rounded overflow-hidden">
-					<div
-						className="progress-done"
-						style={{
-							opacity: 0.75,
-							width: `0%`,
-						}}
-					></div>
-				</div>
-			</div>
-
-			<div className="dark:text-white text-black pl-4 h-20 flex w-full md:w-auto items-center justify-between hover:">
-				<div>
-					<div className="h-4 w-12 bg-gray-400 rounded mb-2"></div>
-					<div className="h-6 w-8 bg-gray-400 rounded"></div>
-				</div>
-			</div>
-		</li>
-	);
-}
+const rankToNumber = (rank: number) => {
+	switch (rank) {
+		case 1:
+			return {
+				color: "#FFD700",
+				fontWeight: 1000,
+			};
+		case 2:
+			return {
+				color: "#C0C0C0",
+				fontWeight: 700,
+			};
+		case 3:
+			return {
+				color: "#CD7F32",
+				fontWeight: 500,
+			};
+		default:
+			return {};
+	}
+};
 
 export default function LeaderboardChild({
 	rank,
@@ -42,34 +34,14 @@ export default function LeaderboardChild({
 	avatarURL,
 }: LeaderboardRankingsInterface) {
 	const [avatar, setAvatar] = useState(avatarURL);
-	let topUsersStyle = {};
-	const rankToNumber = Number(rank);
 
-	switch (rankToNumber) {
-		case 1:
-			topUsersStyle = {
-				color: "#FFD700",
-				fontWeight: 1000,
-			};
-			break;
-		case 2:
-			topUsersStyle = {
-				color: "#C0C0C0",
-				fontWeight: 700,
-			};
-			break;
-		case 3:
-			topUsersStyle = {
-				color: "#CD7F32",
-				fontWeight: 500,
-			};
-			break;
-		default:
-			topUsersStyle = {};
-			break;
-	}
+	const rankNumber = Number(rank);
+	const topUsersStyle = rankToNumber(rankNumber);
 
+	const discriminatorAvatar = Number(discriminator) % 5;
 	const percent = (xp / (level * level * 100)) * 100;
+	const truncatedDiscriminator = ("0000" + discriminator).slice(-4);
+	const messagesToNextLevel = Math.round((level * level * 100 - xp) / 46);
 
 	return (
 		<li className="transition duration-300 ease-in-out dark:bg-zinc-900 bg-zinc-300 flex flex-wrap items-center w-full my-4 p-2 px-4 rounded-lg group hover:bg-zinc-700 dark:hover:bg-zinc-800 shadow overflow-hidden">
@@ -79,23 +51,26 @@ export default function LeaderboardChild({
 						className="transition duration-300 ease-in-out dark:text-zinc-400 dark:group-hover:text-zinc-400 text-zinc-950 group-hover:text-zinc-950 bg-clip-text"
 						style={topUsersStyle}
 					>
-						{rankToNumber}
+						{rankNumber}
 					</span>
 				</div>
 				<Image
 					className="w-16 rounded-full inline-block mx-4 whitespace-nowrap overflow-hidden"
-					src={avatar ?? `https://cdn.discordapp.com/embed/avatars/${Number(discriminator) % 5}.png`}
+					src={avatar ?? `https://cdn.discordapp.com/embed/avatars/${discriminatorAvatar}.png`}
 					alt={`${username}'s Avatar`}
 					style={{ textIndent: "100%" }}
 					width={64}
 					height={64}
-					onError={() => setAvatar(`https://cdn.discordapp.com/embed/avatars/${Number(discriminator) % 5}.png`)}
+					onError={() => setAvatar(`https://cdn.discordapp.com/embed/avatars/${discriminatorAvatar}.png`)}
 				/>
-				<span className="transition duration-300 ease-in-out text-zinc-400 group-hover:text-zinc-400">
+				<span
+					title={username + `#${truncatedDiscriminator}`}
+					className="transition duration-300 ease-in-out text-zinc-400 group-hover:text-zinc-400"
+				>
 					<span className="transition duration-300 ease-in-out dark:text-zinc-200 dark:group-hover:text-white text-zinc-800 group-hover:text-black">
 						{username}
 					</span>
-					<span>#{("0000" + discriminator).slice(-4)}</span>
+					<span>#{truncatedDiscriminator}</span>
 				</span>
 			</div>
